Relicense under MIT/Apache 2.0 dual license
Why relicense?
===

Apache 2.0 is a permissive license, similar to MIT in nature. The
primary difference is that it specifically grants patent rights related
to the code being distributed. This should not matter to our users,
since Basecamp does not (to my knowledge) hold any patents.

However, this does matter when it comes to accepting contributions.
A long accepted idea is "inbound=outbound". This is essentially the idea
that when you contribute to a project that has a certain license, you
are implying that you have the ability to grant that license on the code
you submit. This has recently been made explicit under [GitHub's terms
of service]

By dual licensing under Apache 2.0, we are requiring that all
contributors to the project be able to grant the Apache 2.0 license,
meaning that they must have the ability to grant any patent rights
required to use the code submitted. This means that if someone were to
submit patented code to the project, the liability is on them and not
the Rails project.

[GitHub's terms of service]: https://help.github.com/articles/github-terms-of-service/#6-contributions-under-repository-license

Why not just Apache 2.0?
====

The FSF does not consider Apache 2.0 to be compatible with GPL 2.0. The
MIT license should continue to be offered as an option for those who
wish to use Rails with GPL 2.0 licensed software. Apache 2.0 is
considered compatible with GPL 3.0 and all versions of LGPL.

Why not relicense all of Rails?
===

The steps required to relicense an existing project are poorly defined
and have never been tested in court. The most conservative
interpretation says that to relicense a project, you need the explicit
permission of every person who has ever contributed to the project. This
is not feasible for Rails today. However, at the moment Action text only
has contributors who are employed by the copyright holder, so this
action is straightforward to take for this project today.

What would this mean for Rails?
==

Merging this PR would mean that different parts of the Rails codebase
would be distributed under different licenses. This is inevitably a
somewhat sticky situation, with no great solution. A lot of projects
solve this by adding license headers to each individual project, but I
do not think this makes sense for Rails.

The solution that would make the most sense to me is to have all new
code contributed to Rails after Action Text is merged follow the same
licensing terms. This would allow us to simply state "Any code in Rails
committed prior to [date of merge] is distributed under the MIT license.
Any code comitted after [date of merge] is distributed under either the
MIT or Apache 2.0 licenses at your option".

This effectively means that Rails continues to be an MIT licensed
project, but all new code submitted must explicitly grant patent rights.
For our users, the only folks who would care whether the code they're
using is only MIT or MIT/Apache dual are folks looking for patent troll
protection -- for those folks the outcome is the same as it is today.

Prior Art
===

The Apache license is used by most folks with open source projects and
major patent portfolios, such as Kubernetes (Google) and Swift (Apple).
The MIT/Apache dual license idea comes from the Rust ecosystem, and much
of the wording used in this PR is taken from those projects.